requirements.txt
fastapi
jinja2
python-multipart
sqlalchemy
psycopg2-binary
python-jose
passlib
uvicorn
alembic
apscheduler